Maint/master/stricter deb shipping

+++ tasks/pe_ship.rake
@@ -164,12 +164,15 @@ if Pkg::Config.build_pe
         incoming_dir = args.incoming
         incoming_dir or fail "Adding packages to apt repo requires an incoming directory"
         invoke_task("pl:fetch")
-        Pkg::Util::Net.remote_ssh_cmd(Pkg::Config.apt_host, "/usr/bin/repsimple add_all \
+        stdout, stderr = Pkg::Util::Net.remote_ssh_cmd(Pkg::Config.apt_host, "/usr/bin/repsimple add_all \
             --confdir #{reprepro_confdir} \
             --basedir #{reprepro_basedir} \
             --databasedir #{reprepro_dbdir} \
-            --incomingdir #{incoming_dir}")
+            --incomingdir #{incoming_dir}", true)
 
+        if (stdout + stderr).include?("Skipping inclusion")
+          fail "Unable to add packages to debian repo because it already contains identical files. Perhaps you are trying to ship a deb that already exists. Verify the debs are a newer version than what already exists in #{reprepro_basedir} on #{Pkg::Config.apt_host}"
+        end
         puts "Cleaning up apt repo 'incoming' dir on #{Pkg::Config.apt_host}"
         Pkg::Util::Net.remote_ssh_cmd(Pkg::Config.apt_host, "rm -r #{incoming_dir}")
